Bugfix/drive slow (#12)
* fixes drive slow issue
* fixes math in conversion from body velocity to motor velocity
* changes mbot-upload-firmware to use pinctrl

* remove debug

* Make encoder resolution a float

* Cache the project version and update the build script to organize the release more cleanly

---------

+++ include/config/mbot_classic_config.h
@@ -6,7 +6,7 @@
 
 /* The user can set the encoder resolution to 20, 40, or 48 with a CMake arg. */
 #ifdef USER_ENCODER_RES
-#define ENCODER_RES             USER_ENCODER_RES
+#define ENCODER_RES             (float)USER_ENCODER_RES
 #else
 #define ENCODER_RES             48.0  // Default encoder resolution.
 #endif   /* USER_ENCODER_RES */
